Restructure components for coordinator pattern integration

+++ internal/indexer/indexer.go
@@ -22,7 +22,7 @@ type Indexer interface {
 }
 
 type IndexStats struct {
-	TotalSize        uint64
+	TotalSize        int
 	DocumentsIndexed int
 }
